change the struction of project

+++ src/inference_qa_llama3.py
@@ -3,13 +3,15 @@
 benchmarking MDQA  能够指定正确答案在哪一个部分
 '''
 import os
+os.environ["http_proxy"] = "127.0.0.1:7890"
+os.environ["https_proxy"] = "127.0.0.1:7890"
 import pdb
 import json
 import math
 import torch
 import logging
-import sys
 import warnings
+import pickle
 import argparse
 import dataclasses
 import numpy as np
@@ -17,11 +19,22 @@ from accelerate import Accelerator
 from accelerate.utils import gather_object
 from tqdm import tqdm
 from rouge import Rouge
+import sys
 from xopen import xopen
 import logging
-import pickle
 from copy import deepcopy
+import datetime
+from transformers import AutoModelForCausalLM, AutoTokenizer, AutoConfig
 from utils.lost_in_the_middle.eval_qa_response import evaluate_qa
+from utils.log_utils.utils import get_git_commit_hash
+from huggingface_hub import login
+login(token="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")
+
+def beijing(sec, what):
+    beijing_time = datetime.datetime.now() + datetime.timedelta(hours=8)
+    return beijing_time.timetuple()
+
+logging.Formatter.converter = beijing
 accelerator = Accelerator()
 
 warnings.filterwarnings('ignore')
@@ -35,13 +48,24 @@ from utils.lost_in_the_middle.prompting import (
     get_synthwiki_qa_prompt
 )
 
-from utils.modify_arch.setup_softmax import setup_models_softmax
+from modify_arch.setup_normal import setup_models
 
 def set_seed(args):
     np.random.seed(args.seed)
     torch.manual_seed(args.seed)
     if args.n_gpu > 0:
         torch.cuda.manual_seed_all(args.seed)
+
+def format_chat_prompt(prompt, tokenizer):
+    messages =messages = [
+        {"role": "system", "content": "You are a helpful chatbot!"},
+        {"role": "user", "content": prompt},
+    ]
+    messages_tokenized = tokenizer.apply_chat_template(
+        messages, tokenize=False, add_generation_prompt=True, return_tensors="pt"
+    )
+    return messages_tokenized
+
 
 def format_instruct_prompt(instruction):
     INSTRUCTION_KEY = "### Instruction:"
@@ -72,40 +96,47 @@ if __name__ == '__main__':
     parser.add_argument("--cache_dir", type=str, default=None)
     parser.add_argument("--seed", type=int, default=42, help="random seed for initialization")
 
-    parser.add_argument("--apply_layers", type=str, default="")
+    parser.add_argument('--enable_ms_poe', action='store_true')
     parser.add_argument('--enable_changed_rope', action='store_true')
-    parser.add_argument("--head_type", type=str, default=None)
-    parser.add_argument('--only_true', action='store_true', help='Only use the relevent documenets in the prompt')
+    parser.add_argument("--apply_layers", type=str, default="")
+    parser.add_argument("--context_len", type=int, default=512)
 
+    parser.add_argument('--only_true', action='store_true', help='Only use the relevent documenets in the prompt')
     parser.add_argument("--sample_num", type=int, default=10)
     parser.add_argument("--answer_idx", type=int, default=0)
     parser.add_argument("--batch_size", type=int, default=2)
     args = parser.parse_args()
-    logging_filename = "../log/softmax/result.log"
+
+    commit_hash_id =  get_git_commit_hash()
+    logging_filename = f"../log/llama3/{commit_hash_id}_result.log"
     directory = os.path.dirname(logging_filename)
     os.makedirs(directory, exist_ok=True)
-    with open(logging_filename, 'w') as file:
-        file.write("")
+    if not os.path.exists(logging_filename):
+        with open(logging_filename, 'w') as file:
+            pass
     logger = logging.getLogger(__name__)
     logging.basicConfig(format="%(asctime)s - %(module)s - %(levelname)s - %(message)s", level=logging.INFO, filename=logging_filename, filemode='a')
 
     if accelerator.is_main_process:
-        logger.info("running %s", " ".join(sys.argv))
-    #处理相关的step
+        logger.info(" ".join(["输入文件："+args.input_path,"正确答案的编号："+str(args.answer_idx)]))
+
     ## set up device
     args.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
     args.n_gpu = torch.cuda.device_count()
     set_seed(args)
 
-    ## Loading Models
-    config, tokenizer, model = setup_models_softmax(args)
+    model = AutoModelForCausalLM.from_pretrained(args.model_name).cuda()
+    tokenizer = AutoTokenizer.from_pretrained(args.model_name, use_fast=False, cache_dir=args.cache_dir, padding_side='left')
+    if tokenizer.pad_token is None:
+        tokenizer.pad_token = tokenizer.eos_token
+    config = AutoConfig.from_pretrained(args.model_name)
     accelerator.wait_for_everyone()
 
     ## Loading Dataset
     examples = []
     prompts = []
     all_model_documents = []
-    if not args.input_path.endswith("pickle"):
+    if "pickle" not in args.input_path:
         with xopen(args.input_path, 'r') as f:
             for line in f:
                 if line.strip() != '':
@@ -132,13 +163,13 @@ if __name__ == '__main__':
                             query_aware_contextualization=False,
                             answer_idx=args.answer_idx
                         )
-                    if "instruct" in args.model_name:
+                    if "Instruct" in args.model_name:
                         prompt = format_instruct_prompt(prompt)
+                    # prompt = format_chat_prompt(prompt, tokenizer)
                     prompts.append(prompt)
                     examples.append(deepcopy(input_example))
                     all_model_documents.append(documents)
     else:
-        #synthwiki
         synthwiki_data = pickle.load(open(args.input_path,"rb"))
         for single_data in synthwiki_data:
             question = single_data["question"]
@@ -152,9 +183,8 @@ if __name__ == '__main__':
             prompts.append(prompt)
             examples.append(deepcopy(single_data))
             all_model_documents.append(documents)
-        if "instruct" in args.model_name:
-            prompt = format_instruct_prompt(prompt)
-
+        # if "instruct" in args.model_name:
+        #     prompt = format_instruct_prompt(prompt)
     #对样本进行了采样
     if len(prompts) > args.sample_num:
         prompts = prompts[:args.sample_num]
@@ -168,18 +198,19 @@ if __name__ == '__main__':
         with torch.no_grad():
             for batched_prompts in tqdm(chunks(sub_prompts, args.batch_size), total=math.ceil(len(sub_prompts) / args.batch_size)):
                 if args.batch_size > 1:
-                    input_ids = tokenizer(batched_prompts, add_special_tokens=False, return_tensors='pt', truncation=True, max_length=config.max_position_embeddings, padding=True).input_ids.to(model.device)
+                    tok_inputs = tokenizer(batched_prompts, add_special_tokens=False, return_tensors='pt', truncation=True, max_length=config.max_position_embeddings, padding=True)
                 else:
-                    input_ids = tokenizer(batched_prompts, add_special_tokens=False, return_tensors='pt', truncation=True, max_length=config.max_position_embeddings).input_ids.to(model.device)
-                input_ids = input_ids.cuda()
+                    tok_inputs = tokenizer(batched_prompts, add_special_tokens=False, return_tensors='pt', truncation=True, max_length=config.max_position_embeddings)
+                tok_inputs = {key: value.cuda() for key, value in tok_inputs.items()}
+                input_ids = tok_inputs["input_ids"]
+                # print(input_ids.shape)
                 outputs = model.generate(
-                    input_ids=input_ids,
+                    **tok_inputs,
                     max_length=100 + len(input_ids[0]),
                     use_cache=True,
                     return_dict_in_generate=False,
                     do_sample=False
                 )
-
                 for i, generated_sequence in enumerate(outputs):
                     text = tokenizer.decode(generated_sequence, skip_special_tokens=True, clean_up_tokenization_spaces=True)
                     if input_ids is None:
@@ -192,12 +223,11 @@ if __name__ == '__main__':
                                 clean_up_tokenization_spaces=True,
                             )
                         )
-                    new_text = text[prompt_length:]
+                    new_text = text[prompt_length-10:]
                     responses.append(new_text)
 
         #从所有的设备上搜集数据
         accelerator.wait_for_everyone()  
-        #收集所有的数据
         responses=gather_object(responses)
 
         if accelerator.is_main_process:
@@ -209,15 +239,10 @@ if __name__ == '__main__':
                 for example, model_documents, prompt, response in zip(examples, all_model_documents, prompts, responses):
                     output_example = deepcopy(example)
                     # Add some extra metadata to the output example
-                    output_example["model_prompt"] = prompt
-                    output_example["model_answer"] = response
                     # output_example["model_documents"] = [dataclasses.asdict(document) for document in model_documents]
+                    output_example["model_answer"] = response
+                    output_example["model_prompt"] = prompt
                     output_example["model"] = args.model_name
-                    output_example["model_temperature"] = 0
-                    output_example["model_top_p"] = "None"
-                    output_example["model_prompt_mention_random_ordering"] = False
-                    output_example["model_use_random_ordering"] = False
                     f.write(json.dumps(output_example) + "\n")
 
-            #写入完毕之后  进行测试
             evaluate_qa(args.output_path,None,logger)
